+++ internal/component/pyroscope/java/args.go
@@ -1,6 +1,7 @@
 package java
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/grafana/alloy/internal/component/discovery"
@@ -21,12 +22,23 @@ type ProfilingConfig struct {
 	Alloc      string        `alloy:"alloc,attr,optional"`
 	Lock       string        `alloy:"lock,attr,optional"`
 	CPU        bool          `alloy:"cpu,attr,optional"`
+	Event      string        `alloy:"event,attr,optional"`
+	PerThread  bool          `alloy:"per_thread,attr,optional"`
 }
 
 func (rc *Arguments) UnmarshalAlloy(f func(interface{}) error) error {
 	*rc = defaultArguments()
 	type config Arguments
 	return f((*config)(rc))
+}
+
+func (arg *Arguments) Validate() error {
+	switch arg.ProfilingConfig.Event {
+	case "itimer", "cpu", "wall":
+		return nil
+	default:
+		return fmt.Errorf("invalid event: '%s'. Event must be one of 'itimer', 'cpu' or 'wall'", arg.ProfilingConfig.Event)
+	}
 }
 
 func defaultArguments() Arguments {
@@ -38,6 +50,8 @@ func defaultArguments() Arguments {
 			Alloc:      "10ms",
 			Lock:       "512k",
 			CPU:        true,
+			Event:      "itimer",
+			PerThread:  false,
 		},
 	}
 }