+++ CHANGELOG.md
@@ -39,6 +39,8 @@ Main (unreleased)
 - (_Public preview_) Added rate limiting of cluster state changes to reduce the
   number of unnecessary, intermediate state updates. (@thampiotr)
 
+- Allow setting the CPU profiling event for Java Async Profiler in `pyroscope.java` component (@slbucur)
+
 ### Bugfixes
 
 - Fixed a clustering mode issue where a fatal startup failure of the clustering service